FinalProject/create login and register for user
