+++ src/pages/error/PageNotFound/PageNotFoundPage.module.css
@@ -1,0 +1,23 @@
+@charset "UTF-8";
+#page_not_found {
+  text-align: center;
+}
+#page_not_found h1 {
+  margin-top: 30px;
+  font-size: 65px;
+  margin-bottom: 20px;
+}
+
+#page_not_found {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 100vh; /* Đảm bảo phần tử tràn cả chiều cao của màn hình */
+  text-align: center;
+}
+
+#page_not_found img {
+  width: 300px;
+  margin-bottom: 20px;
+}/*# sourceMappingURL=PageNotFoundPage.module.css.map */
