--- src/pages/login_and_register/user/Register/RegisterPage.css
@@ -1,100 +1,0 @@
-.form-control:focus {
-    box-shadow: none;
-}
-
-.thumbimg {
-    border: 1px solid #ddd;
-    border-radius: 4px;
-    padding: 5px;
-    width: 130px;
-    transition: 0.2s;
-}
-
-#image-preview,
-#image-preview_imagesforever {
-    height: 64px;
-    border: 1px dotted #b3b3b39e;
-    border-radius: 2px;
-}
-
-#cancel-btn,
-#cancel-btn_imagesforever {
-    position: absolute;
-    top: -5px;
-    right: -9px;
-    width: 20px;
-    cursor: pointer;
-}
-
-#image-container,
-#image-container_imagesforever {
-    position: relative;
-}
-
-#dropbox *,
-#dropbox_imagesforever * {
-    cursor: pointer;
-}
-
-#dropbox,
-#dropbox_imagesforever {
-    cursor: pointer;
-    background: #f8f8f8;
-    border-radius: 10px;
-    color: dimgray;
-    padding: 5px 5px;
-    position: relative;
-    cursor: pointer;
-}
-
-#image,
-#image_imagesforever {
-    opacity: 0;
-    left: 0px;
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    cursor: pointer;
-}
-
-#dropbox:hover,
-#dropbox_imagesforever:hover {
-    background: #E8F5E9;
-}
-
-#dropbox p,
-#dropbox_imagesforever p {
-    text-align: center;
-    padding: 5px 0;
-    margin: 0px;
-    line-height: 100%;
-    align-content: center;
-    justify-content: center;
-    align-items: center;
-    display: flex;
-    color: #64bbf0;
-}
-
-#dropbox p i,
-#dropbox_imagesforever p i {
-    margin: 0px 10px;
-}
-
-#big_register {
-    width: 100vw;
-    height: 100vh;
-    background: #007bff;
-    background: linear-gradient(to right, #0062E6, #33AEFF);
-}
-
-.card-img-left {
-    width: 45%;
-    /* background: scroll center url("{{asset('Blog/image/register.jfif')}}"); */
-    background-size: cover;
-}
-
-.btn-login {
-    font-size: 0.9rem;
-    letter-spacing: 0.05rem;
-    padding: 0.75rem 1rem;
-}
