+++ src/pages/login_and_register/user/Register/UserRegisterPage.scss
@@ -1,0 +1,129 @@
+:root {
+  --mainColor: #96dfe2;
+  --PrimaryColor: hsl(199, 100%, 33%);
+  --SecondaryColor: hsl(187, 85%, 43%);
+  --gradientColor: linear-gradient(
+    to right,
+    hsl(187, 85%, 43%),
+    hsl(199, 100%, 33%)
+  );
+  --whiteColor: hsl(0, 0%, 100%);
+  --blackColor: hsl(201, 33%, 16%);
+  --textColor: hsl(240, 4%, 36%);
+  --whiteColorDeam: hsl(0, 0%, 93%);
+  --greyText: rgb(190, 190, 190);
+  --inputColor: rgb(239, 239, 239);
+  --bodyColor: rgb(240, 240, 246);
+  --cardBG: rgb(225, 225, 225);
+}
+
+.form-control:focus {
+  box-shadow: none;
+}
+
+.thumbimg {
+  border: 1px solid #ddd;
+  border-radius: 4px;
+  padding: 5px;
+  width: 130px;
+  transition: 0.2s;
+}
+
+#image-preview,
+#image-preview_imagesforever {
+  height: 64px;
+  border: 1px dotted #b3b3b39e;
+  border-radius: 2px;
+}
+
+#cancel-btn,
+#cancel-btn_imagesforever {
+  position: absolute;
+  top: -5px;
+  right: -9px;
+  width: 20px;
+  cursor: pointer;
+}
+
+#image-container,
+#image-container_imagesforever {
+  position: relative;
+}
+
+#dropbox *,
+#dropbox_imagesforever * {
+  cursor: pointer;
+}
+
+#dropbox,
+#dropbox_imagesforever {
+  cursor: pointer;
+  background: #f8f8f8;
+  border-radius: 10px;
+  color: dimgray;
+  padding: 5px 5px;
+  position: relative;
+  cursor: pointer;
+}
+.register-input {
+  border: 1px var(--greyText) solid;
+  border-right: none;
+  border-top: none;
+  border-left: none;
+  padding: 10px 5px;
+  width: 100%;
+}
+#image,
+#image_imagesforever {
+  opacity: 0;
+  left: 0px;
+  width: 100%;
+  height: 100%;
+  position: absolute;
+  cursor: pointer;
+}
+
+#dropbox:hover,
+#dropbox_imagesforever:hover {
+  background: #e8f5e9;
+}
+
+#dropbox,
+#dropbox_imagesforever {
+  p {
+    text-align: center;
+    padding: 5px 0;
+    margin: 0px;
+    line-height: 100%;
+    align-content: center;
+    justify-content: center;
+    align-items: center;
+    display: flex;
+    color: #64bbf0;
+        i{
+            margin: 0px 10px;
+        }
+    }
+}
+
+#big_register {
+  width: 100vw;
+  height: 100vh;
+  background: #96dfe2;
+  /* background: linear-gradient(to right, #0062E6, #33AEFF); */
+}
+
+.card-img-left {
+  width: 45%;
+  /* background: scroll center url("{{asset('Blog/image/register.jfif')}}"); */
+  background-size: cover;
+}
+
+.btn-login {
+  font-size: 0.9rem;
+  letter-spacing: 0.05rem;
+  padding: 0.75rem 1rem;
+}
+.body-register {
+  border-radius: 20px;
+}
